Change touching many files: hash codes improved/changed to unsigned int.
Issues remaining: search "TODO HASH".  Some tests failing, I think due
to changed order of display.  A test file for new hash codes
is yet to be committed.

Also changed: engine objects in hash.hpp.  Hash codes are now only
computed when asked for.

+++ M2/Macaulay2/e/aring-CC.hpp
@@ -44,10 +44,10 @@ namespace M2 {
 
     const ARingRR& real_ring() const { return mRR; }
 
-    unsigned long computeHashValue(const elem& a) const 
+    unsigned int computeHashValue(const elem& a) const 
     { 
       double v = 12347. * a.re + 865800. * a.im;
-      return static_cast<unsigned long>(v);
+      return static_cast<unsigned int>(v);
     }
 
     /////////////////////////////////
