+++ M2/Macaulay2/e/aring-RRR.hpp
@@ -32,10 +32,10 @@ namespace M2 {
     unsigned long get_precision() const { return mPrecision; }
     void text_out(buffer &o) const;
 
-    unsigned long computeHashValue(const elem& a) const 
+    unsigned int computeHashValue(const elem& a) const 
     { 
       double d = mpfr_get_d(&a, GMP_RNDN);
-      return static_cast<unsigned long>(d);
+      return static_cast<unsigned int>(d);
     }
 
     /////////////////////////////////
