+++ M2/Macaulay2/e/aring-qq-flint.hpp
@@ -37,11 +37,11 @@ namespace M2 {
     
     size_t cardinality() const { return static_cast<size_t>(-1); }
 
-    unsigned long computeHashValue(const ElementType& a) const 
+    unsigned int computeHashValue(const ElementType& a) const 
     { 
       unsigned long numhash = fmpz_get_ui(fmpq_numref(&a));
       unsigned long denhash = fmpz_get_ui(fmpq_denref(&a));
-      return 13253 * numhash + 7647 * denhash;
+      return static_cast<unsigned int>(13253 * numhash + 7647 * denhash);
     }
     
     /** @name properties
