+++ M2/Macaulay2/e/aring-tower.hpp
@@ -73,9 +73,9 @@ namespace M2 {
     // The following are he routines required to be of type "RingInterface"
     ///////////////////////////////////
 
-    unsigned long computeHashValue(const elem& a) const 
+    unsigned int computeHashValue(const elem& a) const 
     { 
-      return static_cast<unsigned long>(a->deg);
+      return a->deg;
     }
 
     void text_out(buffer &o) const;
